+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix freezing when running actions.
 - Add action to clear REPL output. #78
 - Add action to refresh all namespaces. #80
+- Add action to refresh changed namespaces. #81
 
 ## 1.3.1